--- src/view/xdg-shell/xdg-toplevel.cpp
@@ -1,7 +1,0 @@
-#include "xdg-toplevel.hpp"
-
-void wf::xdg_toplevel_t::commit()
-{}
-
-void wf::xdg_toplevel_t::apply()
-{}
